feat(inference): one table for per-galaxy catalog channels (#1469)
Catalog writes per-galaxy data onto a dict; three engines read it --
sequential (map, the default), vmapped MCMC, batched native VI. Which
channels each engine read was hand-written per cell, and the default for
an unwritten cell was SILENCE: an engine that never reads a key raises
nothing, it just scores every galaxy against the template Observation.

    channel            native VI   MCMC vmapped   sequential
    flux_obs/noise         y            y             y
    presence            raised          y             y
    redshift            raised          y             y
    line_flux_obs       silent          y          silent (#1599)
    line_censor            -            -             -

Six channels, three engines, four different transport mechanisms, and no
single place enumerating them. #1460, #1480 and #1599 are three
instances of that one shape.

GALAXY_CHANNELS is now the one declaration -- (name, keys, engines,
what, remedy) -- and _refuse_unsupported_channels the one rule,
replacing three hand-written branches. The load-bearing part is the
unknown-key arm: a galaxy-dict key no row declares is refused on every
engine, so the producer and the consumers cannot drift apart in silence.
test_catalog_channel_registry.py pins both directions by AST-scanning
catalog.py for galaxy_dict["..."] writes; neuter-checked by injecting an
undeclared key.

Also closes the two remaining #1469 gaps:

* line_censor_cols= on Catalog/ingest_catalog carries per-galaxy
  emission-line limits, the catalog form of #1460's
  Data(lines={'Halpha': (f, e, 'upper')}). censor_cols is the photometric
  band axis and cannot express one; an Observation-level flag applies to
  the whole catalog, and non-detection is per-galaxy. Same convention as
  censor_cols (0/1/-1, booleans refused). Honored by the sequential
  engine; the batched engines refuse it rather than drop it.

* An unknown fit kwarg now names the method instead of the runner.
  cat.fit(lines=...) raised "run_map() got an unexpected keyword
  argument 'lines'" -- a function the caller never mentioned, inside a
  backend they did not choose. check_unknown_kwargs sits beside
  check_capabilities at the dispatch seam, so ForwardModel.fit and
  SEDModel.fit get the same answer; accepted names are read off the live
  runner signature so the advice cannot drift.

Refs #1469, #1599

+++ tests/contract/_line_catalog_fixture.py
@@ -183,7 +183,10 @@ def build_two_galaxy_catalog(
         "halpha_err": line_err,
     }
     if line_censor is not None:
-        table["halpha_limit"] = np.asarray(line_censor, dtype=int)
+        # dtype is NOT coerced: a boolean column must stay boolean so the
+        # ingest guard can refuse it. Casting here would launder exactly the
+        # include-mask/censor-flag confusion the guard exists to catch.
+        table["halpha_limit"] = np.asarray(line_censor)
 
     # Create catalog with line columns
     if n_line_cols is not None and n_line_cols != len(line_names):
